requirements.txt
Django~=2.2
djangorestframework~=3.12
djangorestframework-simplejwt~=4.7
djoser~=2.1
Pillow~=8.3
pytest~=6.2
pytest-pythonpath~=0.7
pytest-django~=4.4
PyJWT~=2.1
requests~=2.26